exam/1/requirements.txt
amqp==5.0.6
APScheduler==3.7.0
baidu-aip==2.2.18.0
billiard==3.6.4.0
celery==5.1.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cycler==0.10.0
Django==2.2.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==2.10
kiwisolver==1.3.1
kombu==5.1.0
matplotlib==3.4.2
minio==7.0.3
mysqlclient==2.0.3
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
prompt-toolkit==3.0.18
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
qcloudsms-py==0.1.4
requests==2.25.1
simpleui==4.0.2
six==1.16.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
